requirements.txt
annotated-types==0.7.0
anyio==4.9.0
autogen-agentchat==0.5.7
autogen-core==0.5.7
autogen-ext==0.5.7
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Deprecated==1.2.18
distro==1.9.0
Flask==3.1.1
flask-cors==6.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonref==1.1.0
MarkupSafe==3.0.2
openai==1.82.0
opentelemetry-api==1.33.1
pillow==11.2.1
protobuf==5.29.4
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
